pyproject.toml
[project]
name = "voiceaibot"
version = "0.1.0"
description = "Darwin Gödel Machine-inspired debt collection voice agent"
requires-python = ">=3.11"
dependencies = [
    "pipecat-ai[silero,google,deepgram,cartesia]>=0.0.47",
    "google-genai>=1.0.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.9.0",
    "click>=8.1.7",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
